+++ .gitignore
@@ -5,3 +5,4 @@
 /dist
 /installer/dist
 /installer/build
+main.spec
