--- PyQt6 Starter.py
@@ -1,27 +1,0 @@
-#stdlib imports
-import sys
-#external imports
-from PyQt6.QtWidgets import *
-from PyQt6.QtGui import *
-from PyQt6.QtCore import *
-#local imports
-#none!
-
-class MainWindow(QMainWindow):
-    def __init__(self):
-        super().__init__()
-        self.setWindowTitle("PyQt6 Starter")
-        self.resize(800, 600)
-        self.setCentralWidget(QLabel("Hello World!"))
-        self.show()
-
-if __name__ == "__main__":
-    app = QApplication(sys.argv)
-    window = MainWindow()
-    sys.exit(app.exec())
-
-#ignore this file lol
-
-
-
-
