--- build.ps1
@@ -1,1 +1,0 @@
-pyinstaller  --noconsole --icon=icon.ico main.py
