+++ otherStuff/build.ps1
@@ -1,0 +1,1 @@
+pyinstaller  --noconsole --icon=icon.ico main.py
