--- stylesheet.qss
@@ -1,100 +1,0 @@
-
-QMainWindow{
-    background: #1e1e1e;
-    border-radius: 10px;
-}
-QLabel {
-    color: white;
-    background: transparent;
-    font-family: "Urbanist Medium";
-    font-size: 15px;
-    
-}
-/*222222*/
-QPushButton {
-    background: #1e1e1e;
-    border: 1px solid #ffffff;
-    border-radius: 7px;
-    color: white;
-    padding: 5px;
-    font-family: "Urbanist Medium";
-    font-size: 15px;
-}
-QPushButton:hover {
-    background: #ffffff;
-    color: #1e1e1e;
-}
-QPushButton:pressed {
-    background: #e5e3ff;
-    color: #1e1e1e;
-    font-family: "Urbanist Black";
-}
-QPushButton:disabled {
-    background: #361616;
-    color: #ffffff;
-}
-QProgressBar {
-    border: 1px solid #ffffff;
-    background: #1e1e1e;
-    color: white;
-    text-align: center;
-}
-QProgressBar::chunk {
-    background-color: #ffffff;
-}
-QDialog {
-    background: #1e1e1e;
-    border: 1px solid #ffffff;
-    border-radius: 7px;
-    color: white;
-    padding: 5px;
-}
-
-QWidget{
-    background: #1e1e1e;
-}
-
-QTabWidget::pane {
-    background: #1e1e1e;
-    color: white;
-}
-QTabBar::tab {
-    background: #1e1e1e;
-    color: white;
-    border: 1px solid #ffffff;
-    border-top-left-radius: 7px; /* For rounded corners */
-    border-top-right-radius: 7px;
-    min-width: 8ex;
-    padding: 2px; /* Padding inside each tab */
-    margin-right: 2px;
-    font-family: "Urbanist Light";
-}
-QTabBar::tab:selected {
-    background: #2e2e2e;
-    color: white;
-    border: 2px solid #ffffff;
-    border-top-left-radius: 7px; /* For rounded corners */
-    border-top-right-radius: 7px;
-    min-width: 8ex;
-    padding: 2px; /* Padding inside each tab */
-    margin-right: 2px;
-    font-family: "Urbanist Medium";
-}
-
-QLineEdit {
-    background: #1e1e1e;
-    border: 1px solid #ffffff;
-    color: white;
-    padding: 2px;
-}
-
-QErrorMessage::QTextEdit{
-    color: #ffffff;
-}
-
-QComboBox {
-    color: #ffffff
-
-
-
-}
